Factor XCurveJacobian_ into shared curvejacobian.hpp header
Extracts the dense Jacobian subclass (struct XCurveJacobian_ : Underdetermined::Jacobian_)
from calibration.cpp into a new shared header dal-cpp/dal/curve/curvejacobian.hpp. The
subclass is structurally identical to what the joint multi-curve AAD path needs; factoring
avoids a second copy that must track the first (design §8 OQ-2 resolution).

calibration.cpp drops the local definition and adds #include <dal/curve/curvejacobian.hpp>.
Pure refactor -- no behavior change. All 34 tests in AnalyticJacobianTest + CalibrationTest +
JointCalibrationTest + JointAnalyticJacobianTest pass on the native build.
